backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv

sqlalchemy
psycopg2-binary
alembic

passlib[bcrypt]
python-jose[cryptography]

pydantic[email]
httpx

itsdangerous
